refactor #48: 네이밍 수정
