+++ coupon-service/src/main/java/com/musinsam/couponservice/app/application/dto/v1/coupon/response/ResCouponUseDtoApiV1.java
@@ -1,10 +1,10 @@
 package com.musinsam.couponservice.app.application.dto.v1.coupon.response;
 
 
-import com.musinsam.couponservice.app.doamin.entity.coupon.CouponEntity;
-import com.musinsam.couponservice.app.doamin.entity.coupon.CouponPolicyEntity;
-import com.musinsam.couponservice.app.doamin.vo.CouponStatus;
-import com.musinsam.couponservice.app.doamin.vo.DiscountType;
+import com.musinsam.couponservice.app.domain.entity.coupon.CouponEntity;
+import com.musinsam.couponservice.app.domain.entity.coupon.CouponPolicyEntity;
+import com.musinsam.couponservice.app.domain.vo.CouponStatus;
+import com.musinsam.couponservice.app.domain.vo.DiscountType;
 import java.time.ZonedDateTime;
 import java.util.UUID;
 import lombok.AccessLevel;
@@ -19,7 +19,7 @@ public class ResCouponUseDtoApiV1 {
 
   private Coupon coupon;
 
-  public static ResCouponUseDtoApiV1 from(CouponEntity couponEntity) {
+  public static ResCouponUseDtoApiV1 of(CouponEntity couponEntity) {
     return ResCouponUseDtoApiV1.builder()
         .coupon(Coupon.from(couponEntity))
         .build();
