+++ coupon-service/src/main/java/com/musinsam/couponservice/app/application/dto/v1/couponPolicy/response/ResCouponPolicyIssueDtoApiV1.java
@@ -1,8 +1,8 @@
 package com.musinsam.couponservice.app.application.dto.v1.couponPolicy.response;
 
 
-import com.musinsam.couponservice.app.doamin.entity.coupon.CouponPolicyEntity;
-import com.musinsam.couponservice.app.doamin.vo.DiscountType;
+import com.musinsam.couponservice.app.domain.entity.coupon.CouponPolicyEntity;
+import com.musinsam.couponservice.app.domain.vo.DiscountType;
 import java.time.ZonedDateTime;
 import java.util.UUID;
 import lombok.AccessLevel;
